requirements.txt
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.1
pydantic>=2.5
python-dotenv>=1.0
tenacity>=8.2
rich>=13.7
pytest>=8.0
